Fix issue where ItemDetailView has white navigation bar in split view menu on iPad. Move ios 13 navigation bar code to extension for reuse
